Use next_node as fallback in multiple choice question
Instead of the next_node command overriding the other calculations, use
it as a fallback if no other match is found. I haven't verified that no
flows rely upon this behaviour but am in the process of going through
every flow to rewrite the next_node calculations. I would be surprised
if the override semantics would have been relied upon as it's a slightly
counter-intuitive behaviour anyway (I think treating it as a default
would be more of the expected behaviour).