Add checkbox to child maintenance calculator
The `/calculate-your-child-maintenance/y/pay/1_child` needs an additional checkbox `None of the above` to handle the `no` option.

Currently the user can leave the options blank, or select the `None of the above` checkbox which takes them to the `What is your weekly gross income?` step. Any other checkboxes selected takes them to the `/skillseekers_training` step.