[WIP] Script to update marriage abroad outcomes
A rough and half-ready script to bulk update `_opposite_sex.govspeak.erb` files
for each country.

The content team propose 4 variations of text to add:
1 - Adding text including a link to the end of the file
2 - Adding text without a link to the end of the file
3 - Inserting text above `##Prove you’re free to get married`
4 - Inserting text above `##What you need to do`

This can be specified in the csv file, which the script uses to determine
how to update the various files for each country.

Variations 1 and 2 have been done to test it works (will need some tweaking!),
3 and 4 follows the same method except will need to use something like `gsub`
to insert the text in a particular place in the markdown.

It also needs to use a regex to accommodate countries that have a `the`,
for example `The Falkland Islands`, where it's stripped to match the
directory which is `falkland-islands`, but shouldn't strip it from countries
like `Democratic Republic of the Congo`.

Run with `rails r script/update-marriage-abroad-opposite-sex-outcomes.rb`

When running the script we can pass additional arguments to decide whether to
bulk update or update by country (updating by country means we can separate the
changes in commits and will be easier to review).
As we update each country we can update an additional column in the CSV which
be checked if the script needs to be run multiple times. This will allow us to
skip over countries that have already been updated.

Writing a script to remove the updates should be fairly straightforward to
reverse using something like `gsub` and a regex.