Quick explore of more flexible content options
This is a quick proof of concept for an idea that could allow for
options to embed HTML into Smart Answers to allow greater flexibility
for customisation.

A couple of things this would likely need are:

- removing the html argument used when calling content_for on the
  ErbRenderer
- applying the stripping and line break removal to govspeak and text
- dropping the .govspeak part of erb filenames to reflect their wider
  usage
- a whole bunch of edits to existing Smart Answers and documentation

There's also the ability to render most of the views as text by
appending .text on the URL. This doesn't work on GOV.UK so not sure what
it is actually for, but embedding HTML may cause problems for it.