Updates to Q1 - Q4
It's not possible to access the calculator outside the blocks in
the calculator flow, which means we can't make questions dynamic
(where the options are based on previous responses from the user)

For Q2 `:which_tax_year?`, as the tax year options are constant and
are extracted form the yaml file we can directly call the call method
(rather than storing it in a class variable).

For Q3 `:how_many_children_part_year?` question we have now changed it
to use the `value_question`, and validate the user input in the
`ChildBenefitTaxCalculator` class.

For Q4 `:income_details?` we have started copying over the calculations
from the `Calculators` app and added some validation to the question
flow.