Change Q1 to be a Value question
This removes the use of radio buttons and requires the user to
enter the number of children.
We have limited the max number of children to be `30`, which was
the original number when this calculator was first created.

The validation method for the `how_many_children_part_year?` has
been renamed to `valid_number_of_part_year_children` which is clearer,
and `valid_number_of_children` can now be used for Q1.