Add Results outcome page
We have copied all the copy over from Calculators (calculators/app/views/child_benefit_tax/_results.html.erb)
and updated the flow to route to this outcome page.

We updated Q5 and Q6 to use the `Money` question so that we can
use it's built in validation.

We also added some methods used in the Calculators helper file
(app/helpers/child_benefit_tax_helper.rb) which are used in the
results view.