Prevent self-employed mothers from answering unnecessary earnings question
If a self-employed mother has answered "yes" to working 26 weeks there is
no need to also ask if they have earned £390. This change skips this question in this case.

It also adjusts a method on the calculator to not examine the unanswered question
for self-employed mothers, as this can be used in later calculations if the
mother has a partner.