Don't use `normalize_responses_param` to get session response
As mentioned in the previous commit, `normalize_responses_param` and therefore
`all_responses` must return an array, therefore we need to find another way to get
the responses for session-flows. We were only using `normalize_responses_param` to get
`params[:responses]` so it's just a case of calling that directly.