Adds quick escape banner
This changes the design completely, from a link in the bottom of the page to a red button in a sticky banner at the top.

A z-index is also applied to ensure that the banner is always on the top.

A sticky progressive enhancement is applied via JS to add sticky banner support for IE11.

The button styling has been overridden to add bold styling to match designs and give better presense on the page.