Adds onboarding message
This change adds the onboarding message to the banner. The onboarding message allows users to understand what the button will do on each app.

This is passed in as a variable as the button may behave differently according the app set up.

There are additional styling changes made, such as padding and a fixed with for the button grid column to ensure that the styling stays consistent regardless of screensize.

A non standard size of govuk-spacing(1) + 3 has been used for the padding top for the information section to ensure that the details component sits visually aligned to the middle of the button.